Update remove to add a debug error to test UI

+++ backend/src/notes/notes.service.ts
@@ -1,10 +1,12 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { Injectable, NotFoundException, RequestTimeoutException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Connection, Repository } from 'typeorm';
 import { PaginationQueryDto } from 'src/common/dto/pagination-query.dto';
 import { CreateNoteDto } from './create-note.dto';
 import { Note } from './note.entity';
 import { UpdateNoteDto } from './update-note.dto';
+import { HttpErrorByCode } from '@nestjs/common/utils/http-error-by-code.util';
+
 
 @Injectable()
 export class NotesService {
@@ -51,8 +53,23 @@ export class NotesService {
     return this.noteRepository.save(note);
   }
 
+  PromiseTimeout(delayms) {
+    return new Promise(function (resolve, reject) {
+        setTimeout(resolve, delayms);
+    });
+}
+
   async remove(id: number) {
+    console.log('remove', id);
+    await this.PromiseTimeout(4000); // For debug
+    console.log('after delay');
+    throw new RequestTimeoutException(`There is an error in removing note`);
+    
     const note = await this.noteRepository.findOne(id);
+    if (!note) {
+      throw new NotFoundException(`Note #{id} not found`);
+    }
+    
     return this.noteRepository.remove(note);
   }
 }
